Split out unit/integration tests into separate files
